% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{focal_species}
\alias{focal_species}
\title{Focal species names and publication history}
\format{A data frame with 29 rows and 14 variables:
\describe{
  \item{prior_code}{four letter code, first two of genus and first two of species, used in previous publications}
  \item{prior_name}{prior scientific name}
  \item{prior_native}{prior classification as native or non-native}
  \item{current_code}{revised code}
  \item{current_name}{revised scientific name}
  \item{current_native}{revised classification as native or non-native}
  \item{family}{family}
  \item{reason_updated}{notes on revision}
  \item{Levine_HilleRisLambers_2009_Nature}{included in this paper}
  \item{Godoy_Levine_2014_Ecology}{included in this paper}
  \item{Godoy_etal_2014_EcolLett}{included in this paper}
  \item{Kraft_etal_2015_PNAS}{included in this paper}
  \item{Godoy_etal_2017_Ecology}{included in this paper}
  \item{Saavedra_etal_2017_EcolMono}{included in this paper}
}}
\source{
\url{https://github.com/wpetry/SedgwickTaxonomy}
}
\usage{
focal_species
}
\description{
A dataframe with all the annual plant species used in experiments by
Levine, HilleRisLambers, Godoy, Kraft and others. Includes past names used
and currently revised names. Dataframe is taken directly from the
SedgwickTaxonomy project by Will Petry.
}
\keyword{datasets}
