% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sedgwick_plants}
\alias{sedgwick_plants}
\title{Sedgwick species names}
\format{A data frame with 733 rows and 16 variables:
\describe{
  \item{calflora_taxon}{scientific name from Calflora}
  \item{common}{common name}
  \item{life_history}{annual or perennial}
  \item{form}{growth form, herb or grass etc.}
  \item{status}{native or non-native}
  \item{family}{family}
  \item{calflora_binomial}{scientific binomial, genus, species from Calflora}
  \item{synonym}{hand coded synonym for taxa for which the Calflora name is not widely recognized}
  \item{standard_binomial}{standard scientific binomial}
  \item{genus}{genus from standard binomial}
  \item{species}{species from standard binomial}
  \item{prior_name}{prior name used by Levine et al.}
  \item{current_code}{current code used by Levine et al.}
  \item{prior_code}{prior code used by Levine et al.}
  \item{focal_species}{Logical, whether this was used as a focal species in previous research by Levine et al.}
  \item{USDA_symbol}{Standard code from the USDA plants database. Codes for the species binomial are used rather than codes for a particular subspecies or variety}
}}
\source{
\url{http://www.calflora.org/entry/wgh.html#srch=t&group=none&fmt=photo&inma=t&y=34.6927&x=-120.042&z=12&lpstr=t&lpom=d}
}
\usage{
sedgwick_plants
}
\description{
A dataframe with all the species recorded at Sedgwick based on the Calflora
species occurence data. Also flags focal whether species are focal species
in the experiments conducted by Levine, HilleRisLambers, Godoy, Kraft and
others.
}
\keyword{datasets}
